Simplify SD-card Logger to write when a block is full
Change the previous Logger strategy from buffering many LogEntries at
once and then writing the affected blocks to the SD-card after a
timeout or once the buffer is full. The new strategy uses a smaller
buffer (2 x block size = 1024 byted) and simplified logic to always
write a block once it is full and move the remaining bytes to the
beginning of the buffer for the next block.

This simplifies the logic and removes previously error-prone slice
management and the use of a `static mut` as a buffer (no more use of
`unsafe`).