Move existing code into new firmware folder
The project is growing and I want to refactor the code that does the
serialization of recordings/logs into its own crate to facilitate the
creation of a CLI to deseralize the data into JSON and/or CSV files.
Therefore all code beloning to the actual hardware was moved into
a new firmware folder.